Split the code into separate files for each class
Rubocop issued warnings, so I resolved it by splitting each class into its own file.

```
lib/delayed/backend/active_record.rb:33:7: C: Metrics/BlockLength: Block has too many lines. [122/25]
      ActiveSupport.on_load(:active_record) do ...
      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
lib/delayed/backend/active_record.rb:36:9: W: Lint/ConstantDefinitionInBlock: Do not define constants this way within a block.
        class Job < ::ActiveRecord::Base ...
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
```
